new plot and performance class + tidy do_detection() #226
