+++ algorithms/main_example.py
@@ -5,8 +5,7 @@ from algorithms.decision_engines.stide import Stide
 from algorithms.ids import IDS
 from dataloader.data_loader import DataLoader
 from dataloader.data_preprocessor import DataPreprocessor
-import pprint
-from score_plot import ScorePlot
+from pprint import pprint
 
 
 if __name__ == '__main__':
@@ -37,10 +36,8 @@ if __name__ == '__main__':
     ids.train_decision_engine()
     ids.determine_threshold()
     ids.do_detection()
-    pprint.pprint(ids.get_performance())
+    pprint(ids.performance.get_performance())
 
-    # creating plot
-    plot = ScorePlot(scenario_path=dataloader.scenario_path)
-
-    plot.feed_figure(ids.get_plotting_data())
-    plot.show_plot()
+    # plotting
+    ids.plot.feed_figure()
+    ids.plot.show_plot()
